Add use case where receiver task breaks
The error is caused as follows:
A future is created in the while loop of retry_get_event_sample().
When the retry_get_event_sample task gets timeout the future is cancelled:
"If a timeout occurs, it cancels the task and raises asyncio.TimeoutError."
Therefore the future is set to None which is why we get an
InvalidStateError when the receiver task tries to set it to some value.